refactor compression

+++ iotdb-protocol/thrift-datanode/src/main/thrift/client.thrift
@@ -243,8 +243,8 @@ struct TSInsertTabletReq {
   9: optional bool writeToTable
   10: optional list<byte> columnCategories
   11: optional bool isCompressed
-  12: optional list<i32> encodingTypes
-  13: optional i32 compressType
+  12: optional list<byte> encodingTypes
+  13: optional byte compressType
 }
 
 struct TSInsertTabletsReq {
